Data Explorer save as overlay uses router
