+++ ui/src/pageLayout/components/Page.tsx
@@ -11,7 +11,6 @@ import PageContents from 'src/pageLayout/components/PageContents'
 import {ErrorHandling} from 'src/shared/decorators/errors'
 
 interface Props {
-  children: JSX.Element | JSX.Element[]
   className?: string
   titleTag: string
 }
